remove wikipedia irc firehose and dependencies from core server module to examples
